+++ common/src/AY-3-8910.c
@@ -22,4 +22,4 @@ void stopSounds(void) {}
 
 void soundTick(void) {}
 
-void playSound(int ununused) {}
+void playSound(uint8_t ununused) {}
