+++ common/src/Globals.c
@@ -2,5 +2,6 @@
  Created by Daniel Monteiro on 03/10/2019.
 */
 
+#include <stdint.h>
 #include "Globals.h"
-int unused = 0;
+uint8_t unused = 0;
